refactor(schedule-details): consolidate cron helper types into hook types
Moves each helper's params/result types out of per-helper .types.ts
files and into the parent use-schedule-runs-chart-data.types.ts, since
they only serve internal helpers of that hook.
