+++ src/views/schedule-details/hooks/use-schedule-runs-chart-data/helpers/get-expected-schedule-times-ms.ts
@@ -2,8 +2,9 @@ import { CronExpressionParser } from 'cron-parser';
 
 import { cronValidate } from '@/utils/cron-validate/cron-validate';
 
+import { type GetExpectedScheduleTimesMsParams } from '../use-schedule-runs-chart-data.types';
+
 import { MAX_SCHEDULE_CRON_OCCURRENCES } from './get-expected-schedule-times-ms.constants';
-import { type GetExpectedScheduleTimesMsParams } from './get-expected-schedule-times-ms.types';
 
 export default function getExpectedScheduleTimesMs({
   cronExpression,
