--- src/views/schedule-details/hooks/use-schedule-runs-chart-data/helpers/get-expected-schedule-times-ms.types.ts
@@ -1,6 +1,0 @@
-export type GetExpectedScheduleTimesMsParams = {
-  cronExpression: string;
-  startMs: number;
-  endMs: number;
-  limit?: number;
-};
