--- src/views/schedule-details/hooks/use-schedule-runs-chart-data/helpers/get-schedule-timeline-bounds.types.ts
@@ -1,12 +1,0 @@
-import { type DescribeScheduleResponse } from '@/route-handlers/describe-schedule/describe-schedule.types';
-
-export type GetScheduleTimelineBoundsParams = {
-  describeSchedule: DescribeScheduleResponse | undefined;
-  retentionSeconds: number | null;
-  nowMs: number;
-};
-
-export type ScheduleTimelineBounds = {
-  inferenceStartMs: number | null;
-  scheduleEndMs: number | null;
-};
