--- src/views/schedule-details/hooks/use-schedule-runs-chart-data/helpers/get-skipped-schedule-executions.types.ts
@@ -1,10 +1,0 @@
-export type GetSkippedScheduleExecutionsParams = {
-  cronExpression: string;
-  inferenceStartMs: number | null;
-  scheduleEndMs: number | null;
-  oldestLoadedScheduleTimeMs: number | null;
-  hasNextPage: boolean;
-  nowMs: number;
-  nextExecutionTimeMs?: number | null;
-  actualTimesMs: number[];
-};
