--- src/views/schedule-details/hooks/use-schedule-runs-chart-data/helpers/get-skipped-schedule-times-ms.types.ts
@@ -1,4 +1,0 @@
-export type GetSkippedScheduleTimesMsParams = {
-  expectedTimesMs: number[];
-  actualTimesMs: number[];
-};
