+++ src/views/schedule-details/hooks/use-schedule-runs-chart-data/use-schedule-runs-chart-data.types.ts
@@ -1,3 +1,4 @@
+import { type DescribeScheduleResponse } from '@/route-handlers/describe-schedule/describe-schedule.types';
 import { type ChartSeriesData } from '@/views/schedule-details/schedule-details-runs-chart-series/schedule-details-runs-chart-series.types';
 
 export type UseScheduleRunsChartDataParams = {
@@ -10,4 +11,38 @@ export type UseScheduleRunsChartDataParams = {
 export type UseScheduleRunsChartDataResult = {
   data: ChartSeriesData;
   isLoading: boolean;
+};
+
+export type GetScheduleTimelineBoundsParams = {
+  describeSchedule: DescribeScheduleResponse | undefined;
+  retentionSeconds: number | null;
+  nowMs: number;
+};
+
+export type ScheduleTimelineBounds = {
+  inferenceStartMs: number | null;
+  scheduleEndMs: number | null;
+};
+
+export type GetExpectedScheduleTimesMsParams = {
+  cronExpression: string;
+  startMs: number;
+  endMs: number;
+  limit?: number;
+};
+
+export type GetSkippedScheduleTimesMsParams = {
+  expectedTimesMs: number[];
+  actualTimesMs: number[];
+};
+
+export type GetSkippedScheduleExecutionsParams = {
+  cronExpression: string;
+  inferenceStartMs: number | null;
+  scheduleEndMs: number | null;
+  oldestLoadedScheduleTimeMs: number | null;
+  hasNextPage: boolean;
+  nowMs: number;
+  nextExecutionTimeMs?: number | null;
+  actualTimesMs: number[];
 };
